feat: DBのAxiosインスタンスを生成する関数を簡略化しました。
