--- src/utils/anime_db/animeAxios.js
@@ -1,5 +1,0 @@
-import createDbAxios from "../createDbAxios";
-
-const animeAxios = createDbAxios("9ec213bc61bd4af5bbdfc74b844cb4b7");
-
-export default animeAxios;
