--- src/utils/anime_db/getAnime.js
@@ -1,12 +1,0 @@
-import animeAxios from "./animeAxios";
-
-const getAnime = async () => {
-  try {
-    const response = await animeAxios.post("/query");
-    return response.data.results;
-  } catch (error) {
-    console.error(error);
-  }
-};
-
-export default getAnime;
