+++ src/utils/createDbAxios.js
@@ -1,8 +1,8 @@
 import axios from "axios";
 import API_KEY from "./env";
 
-const createDbAxios = (databaseId) => {
-  const baseURL = `https://api.notion.com/v1/databases/${databaseId}`;
+const createDbAxios = (path) => {
+  const baseURL = `https://api.notion.com/v1/${path}`;
   const headers = {
     Authorization: `Bearer ${API_KEY}`,
     "Notion-Version": "2022-06-28",
