--- src/utils/reviews_db/getReviews.js
@@ -1,12 +1,0 @@
-import reviewsAxios from "./reviewsAxios";
-
-const getReviews = async () => {
-  try {
-    const response = await reviewsAxios.post("/query");
-    return response.data.results;
-  } catch (error) {
-    console.error(error);
-  }
-};
-
-export default getReviews;
