--- src/utils/reviews_db/reviewsAxios.js
@@ -1,5 +1,0 @@
-import createDbAxios from "../createDbAxios";
-
-const reviewsAxios = createDbAxios("d571c5897d854e54bd89275d12338a35");
-
-export default reviewsAxios;
